Revamp the GET state machine to avoid stat.
The original GET implementation always started with a standalone `stat`
call - whose results were only used to determine whether to invoke
the directory listing code.

This changes the GET to try to open the path by default and, if it
is either a directory descriptor (or fails with EISDIR), then try
doing the directory listing.

By avoiding the standalone stat, we significantly reduce the number
of filesystem calls done under high concurrency (as the OFS code will
collapse all the open files into a single file handle).